Created helpers doc and moved functions into it
